added abstract code for the robotarm branch
